Okay, the agree and disagree buttons are now working. I've also added a new column to the petitions.csv file to store the number of comments for each petition. Now just need to make sure each user can agree or disagree only once, as now they can just spam.
Also, need to update the ""Petition"" button so that it clears the input area, making it easier for end user

Also, I just realised I need to change these frames to scrollable frames for each frame, so the test is readable.

(Created 9 test petitions today)
Also, the comments section is updated for the post, but then when I cycle back to it, they're not shown.
Although, they're permanently stored in csv file, so it's not so bad

--AA🟥 10:26 180325